+++ jovo-integrations/jovo-nlu-dialogflow/package.json
@@ -15,7 +15,8 @@
   "author": "jovotech",
   "license": "Apache-2.0",
   "dependencies": {
-    "jovo-core": "^2.2.15"
+    "jovo-core": "^2.2.15",
+    "googleapis": "^42.0.0"
   },
   "devDependencies": {
     "@types/dotenv": "^6.1.1",
