+++ jovo-integrations/jovo-nlu-dialogflow/src/index.ts
@@ -1,3 +1,2 @@
 export * from './DialogflowNlu';
-export * from './modules/DialogflowNluCore';
 export * from './Interfaces';
